requirements.txt
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1
matplotlib
numpy
Pillow

